refactor(theme): use new context api; new useTheme hook
BREAKING CHANGE Theme provider component uses new context api
